Code submission for P4: Project Prototype
This zip file includes the catkin_ws zip file which has all of the libraries needed for the Turtlebot3 simulation to run Gazebo and RViz for navigation and mapping. It also includes the map of the house simulation obtained by navigating the robot using the WASDX keys to change the linear and angular velocity which is required for autonomous navigation of the robot.